Add the other REST actions to students; minor restructuring
I made the model(s) implement the models::Model trait, which requires the model
to implement Decodable and Encodable (rustc_serialize) and provides the methods
parse_str(body: &str) -> Option<Self> and to_str(&self) -> Option<String>.
It requires the model to implement the following
methods:
- find_id(id: usize, conn: &Connection) -> Option<Self>
- find_all(conn: &Connection) -> Vec<Self>
- save(self, id: Option<usize>, conn: &Connection) -> Option<Self>
- delete(id: usize, conn: &Connection) -> Option<()>
that handle the database end of the corresponding REST actions.
Using this trait, I added a generic helper function, accordingly called
parse(req: &Request), that procures the request body and parses the body.
I added another helper function, extract_id(req: &Request) -> Option<usize> that
handles the annoying procedure of getting the Params request extension, getting
the id field and converting it to a usize...

I also added another piece of middleware, RequestBody, that reads the whole
request and stores it in a string (it is also the typemap key to said string).