Add support for related resources (for students)
I added support for an ?include= query string for students::index and
students::show, which fetches the associated base set and normal lendings
(as well as the book records). In order to achieve this, I implemented a
skeleton Book record (+constructor), and added two fields to the Student record:
lent_books and base_sets, which contain an Option<Vec<(String, Book)>> where
the String holds the ISO 8601 representation of the date on which the book has
been lent. Those values are usually None (and being ignored) and will only be
retrieved when necessary. To parse the ?include= query string, I added an
enum --- Includable --- whose variants correspond to relationships that may
be included (currently lendings, baseSets and aliases) and a type alias ---
Includes --- for HashSet<Includable> which can be parsed from a request's
query string using handlers::get_includes. I also had to manually implement
rustc_serialize's Decodable for Student so that the parser doesn't throw an
error when id, lent_books and base_sets are not specified (will be None'd
automatically).

I also added better error reporting and logging using the Optionable<T> trait
whose log_to_option method converts the object into an Option<T> and prints
an error message with timestamp if that Option is None. It is currently
implemented for Option<T> (where the specified error message will be displayed)
and for Result<T,E> where E: Debug (where the specified error message will be
displayed alongside the Debug representation of the error value).