Add support for ?include=aliases
On books::index and books::show, the user can now ask for the book's aliases to
also be shown. To accomplish that, I had to do some minor changes (most was
already in place). I had to add an aliases field to the Book record (an
Option<Vec<Alias>>), for which I had to prototype an Alias type. Furthermore, I
head to implement rustc_serialize::Decodable manually, so that the user doesn't
have to specify aliases and id.