Add basic infrastructure for authentication and authorization
I added routes for creating, deleting, editing schools and for creating and
deleting sessions. Before almost each request, the token is looked up and the
school id is determined. If no token is supplied, a 401 Unauthorized will be
returned. When logging in, the user supplied name (of the school) and password.
The server creates a new token (with secret) and returns it to the user
(the same happens when the user creates a new school). Whenever the
authentication_tokens table is accessed, it will be purged of old (older than
a day) entries. When accessing a route thereafter, the token and secret have
to be provided in the Authorization header (Basic Authorizatation).

TODO: Add full authorization support (WHERE clauses in UPDATE, SELECT queries,
etc)
TODO: Make all handlers use handlers::get_db